# api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "H20 Aware API",
    "description": "API documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/precipitation/lat/{lat}/long/{long}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "long",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/evaporation/lat/{lat}/long/{long}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "long",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/clouds/lat/{lat}/long/{long}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "long",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/humidity/lat/{lat}/long/{long}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lat",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "long",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}